address code review: fix typos, upper-case some comments, rename to SportResults
