+++ cdap-examples/SportResults/resources/nfl-2012.csv
