+++ cdap-examples/pom.xml
@@ -37,7 +37,7 @@
     <module>WordCount</module>
     <module>SparkKMeans</module>
     <module>SparkPageRank</module>
-    <module>Sports</module>
+    <module>SportResults</module>
     <module>StreamConversion</module>
     <module>UserProfiles</module>
     <module>WebAnalytics</module>
